refactor: rename components to kebab-case and add UMD dist
