+++ components/status-tag/demos/no-icon.tsx
